[Refactor] Keep Duration as float; only Repeat Duration stays int
Reverts the Duration-column portion of the previous commit. Repeat
Duration remains an int (0–10000) — its QSpinBox editor, int cast in
extract_execution_params, int type in StepParamsCommitMessage, and
int-valued widget/path_execution_service reads are kept.

Duration restored to:
- QDoubleSpinBox editor (decimals=1, step=0.1).
- step_defaults / group_defaults: "1.0".
- _validate_numeric_field uses the :.1f path; invalid → "0.0".
- float casts everywhere (widget, helpers, path_execution_service).
- StepParamsCommitMessage.duration: float.
- RouteLayerManager.duration Range(0.0, 10000.0, 1.0); _current_params
  returns float.
- Tests updated accordingly; legacy-float-string tolerance test narrowed
  to Repeat Duration only.

+++ device_viewer/models/route.py
@@ -213,7 +213,7 @@ class RouteLayerManager(HasTraits):
     phase_label = Str("")
 
     # path execution properties (mirror protocol grid step defaults)
-    duration = Range(low=0, high=10000, value=1)
+    duration = Range(low=0.0, high=10000.0, value=1.0)
     trail_length = Range(low=1, high=10000, value=1)
 
     # Dynamic upper bound for trail_overlay: overlay cannot exceed trail_length - 1
@@ -254,7 +254,7 @@ class RouteLayerManager(HasTraits):
 
     def _current_params(self) -> dict:
         return {
-            "duration": int(self.duration),
+            "duration": float(self.duration),
             "repetitions": int(self.repetitions),
             "repeat_duration": int(self.repeat_duration),
             "trail_length": int(self.trail_length),
